pyproject.toml
[tool.poetry]
name = "digest-task"
version = "0.1.0"
description = ""
authors = ["samsegomof <wemagonof91@gmail.com>"]
readme = "README.md"
packages = [{include = "digest_task"}]

[tool.poetry.dependencies]
python = "^3.10"
poetry = "^1.5.1"
fastapi = {extras = ["uvicorn"], version = "^0.100.0"}
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
alembic = "^1.11.1"
pydantic-settings = "^2.0.2"
greenlet = "^2.0.2"
uvicorn = "^0.23.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
